SYNCOPE-1545: add more tests
