+++ wa/starter/src/main/java/org/apache/syncope/wa/pac4j/saml/SyncopeWASAML2ClientKeystoreGenerator.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.syncope.wa.pac4j;
+package org.apache.syncope.wa.pac4j.saml;
 
 import org.apache.syncope.common.lib.SyncopeClientException;
 import org.apache.syncope.common.lib.to.SAML2SPKeystoreTO;
