+++ wa/starter/src/test/java/org/apache/syncope/wa/pac4j/saml/SyncopeWASAML2ClientCustomizerTest.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.syncope.wa.pac4j;
+package org.apache.syncope.wa.pac4j.saml;
 
 import org.apache.commons.io.IOUtils;
 import org.apache.syncope.client.lib.SyncopeClient;
