+++ wa/starter/src/test/java/org/apache/syncope/wa/pac4j/saml/SyncopeWASAML2ClientMetadataGeneratorTest.java
@@ -17,36 +17,36 @@
  * under the License.
  */
 
-package org.apache.syncope.wa.pac4j;
+package org.apache.syncope.wa.pac4j.saml;
 
 import org.apache.commons.io.IOUtils;
 import org.apache.syncope.client.lib.SyncopeClient;
+import org.apache.syncope.common.lib.SyncopeClientException;
 import org.apache.syncope.common.lib.to.SAML2SPMetadataTO;
 import org.apache.syncope.common.rest.api.service.SAML2SPMetadataService;
 import org.apache.syncope.wa.WARestClient;
 import org.junit.jupiter.api.Test;
+import org.opensaml.saml.saml2.metadata.EntityDescriptor;
 import org.pac4j.saml.client.SAML2Client;
+import org.pac4j.saml.metadata.SAML2MetadataGenerator;
 import org.springframework.core.io.ClassPathResource;
 
 import javax.ws.rs.core.Response;
 
 import java.io.File;
+import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public class SyncopeWASAML2MetadataResolverTest extends BaseSyncopeWASAML2Client {
-    @Test
-    public void fetchMetadata() throws Exception {
-        SAML2Client client = getSAML2Client();
-        String keystoreFile = File.createTempFile("keystore", "jks").getCanonicalPath();
-        client.getConfiguration().setKeystoreResourceFilepath(keystoreFile);
+public class SyncopeWASAML2ClientMetadataGeneratorTest extends BaseSyncopeWASAML2Client {
+    private static WARestClient getWaRestClient(final Response response) throws IOException {
         WARestClient restClient = mock(WARestClient.class);
-
         SAML2SPMetadataTO metadataTO = new SAML2SPMetadataTO.Builder()
             .owner("Syncope")
             .metadata(IOUtils.toString(new ClassPathResource("sp-metadata.xml").getInputStream(), StandardCharsets.UTF_8))
@@ -54,13 +54,35 @@ public class SyncopeWASAML2MetadataResolverTest extends BaseSyncopeWASAML2Client
 
         SAML2SPMetadataService saml2SPMetadataService = mock(SAML2SPMetadataService.class);
         when(saml2SPMetadataService.get(anyString())).thenReturn(metadataTO);
-        when(saml2SPMetadataService.set(any())).thenReturn(Response.ok().build());
+        when(saml2SPMetadataService.set(any())).thenReturn(response);
 
         SyncopeClient syncopeClient = mock(SyncopeClient.class);
         when(syncopeClient.getService(SAML2SPMetadataService.class)).thenReturn(saml2SPMetadataService);
         when(restClient.getSyncopeClient()).thenReturn(syncopeClient);
+        return restClient;
+    }
 
-        SyncopeWASAML2MetadataResolver resolver = new SyncopeWASAML2MetadataResolver(restClient, client);
-        assertNotNull(resolver.fetchMetadata());
+    @Test
+    public void storeMetadata() throws Exception {
+        SAML2Client client = getSAML2Client();
+        String keystoreFile = File.createTempFile("keystore", "jks").getCanonicalPath();
+        client.getConfiguration().setKeystoreResourceFilepath(keystoreFile);
+
+        SAML2MetadataGenerator generator = new SyncopeWASAML2ClientMetadataGenerator(getWaRestClient(Response.ok().build()), client);
+        EntityDescriptor entityDescriptor = generator.buildEntityDescriptor();
+        String metadata = generator.getMetadata(entityDescriptor);
+        assertNotNull(generator.storeMetadata(metadata, null, false));
+    }
+
+    @Test
+    public void storeMetadataFails() throws Exception {
+        SAML2Client client = getSAML2Client();
+        String keystoreFile = File.createTempFile("keystore", "jks").getCanonicalPath();
+        client.getConfiguration().setKeystoreResourceFilepath(keystoreFile);
+        WARestClient restClient = getWaRestClient(Response.serverError().build());
+        SAML2MetadataGenerator generator = new SyncopeWASAML2ClientMetadataGenerator(restClient, client);
+        EntityDescriptor entityDescriptor = generator.buildEntityDescriptor();
+        String metadata = generator.getMetadata(entityDescriptor);
+        assertThrows(SyncopeClientException.class, () -> generator.storeMetadata(metadata, null, false));
     }
 }
